Helper Function of PWCsurvOS()

Helper Function of PWCsurvOS()

PwcOSInt(x, h01, h02, h12, pw01, pw02, pw12)

Arguments

x

(numeric)
variable of integration.

h01

(numeric vector)
constant transition hazards for 0 to 1 transition.

h02

(numeric vector)
constant transition hazards for 0 to 2 transition.

h12

(numeric vector)
constant transition hazards for 1 to 2 transition.

pw01

(numeric vector)
time intervals for the piecewise constant hazards h01.

pw02

(numeric vector)
time intervals for the piecewise constant hazards h02.

pw12

(numeric vector)
time intervals for the piecewise constant hazards h12.

Value

Numeric results of the integrand used to calculate the OS survival function for piecewise constant transition hazards, see PWCsurvOS.

Examples

PwcOSInt(1:5, c(0.3, 0.5), c(0.5, 0.8), c(0.7, 1), c(0, 4), c(0, 8), c(0, 3))
#> [1] 0.2714512 0.2456192 0.2222455 0.4524187 0.4524187