Simulation of a Large Number of Oncology Clinical Trials

Simulation of a Large Number of Oncology Clinical Trials

Usage

getClinicalTrials(nRep, ..., seed = 1234, datType = "1rowTransition")

Arguments

nRep

(int)
number of simulated trials.

...

parameters transferred to getOneClinicalTrial(), see getOneClinicalTrial() for details.

seed

(int)
random seed used for this simulation.

datType

(string)
possible values are 1rowTransition and 1rowPatient.

Value

This function returns a list with nRep simulated data sets in the format specified by datType. See getDatasetWideFormat()

getOneClinicalTrial() for details.

Examples

transition1 <- exponential_transition(h01 = 1.2, h02 = 1.5, h12 = 1.6)
transition2 <- exponential_transition(h01 = 1, h02 = 1.3, h12 = 1.7)
getClinicalTrials(
  nRep = 10, nPat = c(20, 20), seed = 1234, datType = "1rowTransition",
  transitionByArm = list(transition1, transition2), dropout = list(rate = 0.5, time = 12),
  accrual = list(param = "intensity", value = 7)
)
#> [[1]]
#>    id from   to      entry       exit   entryAct   exitAct    censAct trt
#> 1   1    0    2 0.00000000 0.08322401 0.40747241 0.4906964 43.7187692   1
#> 2   2    0    1 0.00000000 0.11092520 1.55505644 1.6659816  5.8270303   1
#> 3   2    1    2 0.11092520 0.41401354 1.66598164 1.9690700  5.8270303   1
#> 4   3    0 cens 0.00000000 0.11394908 0.56049901 0.6744481  0.6744481   1
#> 5   4    0    1 0.00000000 0.60996013 2.56737283 3.1773330 32.7383865   1
#> 6   4    1    2 0.60996013 0.65535210 3.17733296 3.2227249 32.7383865   1
#> 7   5    0    1 0.00000000 0.29849750 1.11285653 1.4113540  7.8158932   1
#> 8   5    1    2 0.29849750 0.54012913 1.41135403 1.6529857  7.8158932   1
#> 9   6    0    2 0.00000000 0.38503933 0.88820223 1.2732416  2.4454416   1
#> 10  7    0    2 0.00000000 0.13841152 0.45722475 0.5956363 14.7240045   1
#> 11  8    0    2 0.00000000 0.36014181 2.56053100 2.9206728  6.0683211   1
#> 12  9    0    2 0.00000000 0.14819800 0.47541080 0.6236088 14.9838502   1
#> 13 10    0    1 0.00000000 0.25820375 2.57264170 2.8308455 15.7374700   1
#> 14 10    1    2 0.25820375 0.94827474 2.83084545 3.5209164 15.7374700   1
#> 15 11    0    2 0.00000000 0.41866506 0.38308056 0.8017456 32.9316082   1
#> 16 12    0    2 0.00000000 0.24576717 0.37604038 0.6218076 28.0083608   1
#> 17 13    0    2 0.00000000 0.10356288 0.30082144 0.4043843 29.0161494   1
#> 18 14    0    1 0.00000000 0.53708453 1.46166737 1.9987519 54.3068613   1
#> 19 14    1    2 0.53708453 2.16778248 1.99875190 3.6294499 54.3068613   1
#> 20 15    0    2 0.00000000 0.02838643 0.85771158 0.8860980 31.1660821   1
#> 21 16    0    1 0.00000000 0.13726281 0.07633399 0.2135968  0.6255771   1
#> 22 16    1    2 0.13726281 0.53631783 0.21359680 0.6126518  0.6255771   1
#> 23 17    0    2 0.00000000 0.46787743 0.88470695 1.3525844 16.0669470   1
#> 24 18    0 cens 0.00000000 0.25299805 2.12034188 2.3733399  2.3733399   1
#> 25 19    0    2 0.00000000 0.06150124 0.10130493 0.1628062 31.8705582   1
#> 26 20    0    2 0.00000000 0.25964639 1.61450318 1.8741496 10.6055161   1
#> 27 21    0    2 0.00000000 0.17049279 0.88598333 1.0564761 14.9813150   2
#> 28 22    0    2 0.00000000 0.28546700 2.12162676 2.4070938 37.0902155   2
#> 29 23    0    2 0.00000000 0.19199585 1.82546090 2.0174567 27.0357872   2
#> 30 24    0    1 0.00000000 0.42932971 2.83575996 3.2650897 20.0766849   2
#> 31 24    1    2 0.42932971 0.96937540 3.26508967 3.8051354 20.0766849   2
#> 32 25    0    2 0.00000000 0.58836837 0.36648511 0.9548535 28.5300295   2
#> 33 26    0    2 0.00000000 0.36288447 2.52354165 2.8864261 27.2107916   2
#> 34 27    0    1 0.00000000 1.71832281 2.31452397 4.0328468 14.7782358   2
#> 35 27    1    2 1.71832281 1.76487186 4.03284678 4.0793958 14.7782358   2
#> 36 28    0    2 0.00000000 0.37388717 2.34814622 2.7220334 48.9485729   2
#> 37 29    0    1 0.00000000 0.25135188 2.38486475 2.6362166  3.8466779   2
#> 38 29    1 cens 0.25135188 1.46181315 2.63621663 3.8466779  3.8466779   2
#> 39 30    0    1 0.00000000 0.11284689 2.09352063 2.2063675  5.4433953   2
#> 40 30    1    2 0.11284689 0.12605695 2.20636752 2.2195776  5.4433953   2
#> 41 31    0    1 0.00000000 0.03727395 2.80869721 2.8459712 47.7400238   2
#> 42 31    1    2 0.03727395 1.12158600 2.84597116 3.9302832 47.7400238   2
#> 43 32    0    2 0.00000000 0.82560334 1.82629881 2.6519022  4.8448472   2
#> 44 33    0 cens 0.00000000 0.11270925 1.88787038 2.0005796  2.0005796   2
#> 45 34    0    1 0.00000000 1.92908730 1.50959822 3.4386855 17.3309479   2
#> 46 34    1    2 1.92908730 2.51783752 3.43868553 4.0274357 17.3309479   2
#> 47 35    0    2 0.00000000 0.40045473 0.90712520 1.3075799 25.2441477   2
#> 48 36    0    2 0.00000000 2.90316744 2.19387277 5.0970402 26.1772667   2
#> 49 37    0    2 0.00000000 0.20476656 1.50373855 1.7085051  7.1172466   2
#> 50 38    0    2 0.00000000 0.35215896 2.09229110 2.4444501  5.4766211   2
#> 51 39    0    2 0.00000000 0.24397560 0.87904497 1.1230206  5.4405979   2
#> 52 40    0    2 0.00000000 0.37292766 1.15478072 1.5277084  7.1877786   2
#> 
#> [[2]]
#>    id from   to      entry        exit  entryAct   exitAct    censAct trt
#> 1   1    0    2 0.00000000 0.113470713 1.0085671 1.1220379  36.005705   1
#> 2   2    0    1 0.00000000 0.121433016 2.8027380 2.9241711   5.023407   1
#> 3   2    1    2 0.12143302 2.017062206 2.9241711 4.8198003   5.023407   1
#> 4   3    0    2 0.00000000 0.024933029 1.5396650 1.5645981  10.474959   1
#> 5   4    0    2 0.00000000 0.306669722 1.2686682 1.5753379   3.580949   1
#> 6   5    0    1 0.00000000 0.112701305 2.7124765 2.8251778  10.376796   1
#> 7   5    1    2 0.11270131 0.286786701 2.8251778 2.9992632  10.376796   1
#> 8   6    0    2 0.00000000 0.001217678 1.2928096 1.2940273  18.360778   1
#> 9   7    0    2 0.00000000 0.329785056 0.5446452 0.8744302  48.623381   1
#> 10  8    0    2 0.00000000 0.271879916 2.8331690 3.1050489  15.397319   1
#> 11  9    0    2 0.00000000 0.689620883 1.5670156 2.2566364  19.442459   1
#> 12 10    0    2 0.00000000 0.011112007 2.1966165 2.2077285  27.380735   1
#> 13 11    0    1 0.00000000 0.339153434 2.6097762 2.9489296  18.655615   1
#> 14 11    1    2 0.33915343 0.925554987 2.9489296 3.5353312  18.655615   1
#> 15 12    0    1 0.00000000 0.115758661 1.9488914 2.0646501  12.234253   1
#> 16 12    1    2 0.11575866 1.002394306 2.0646501 2.9512857  12.234253   1
#> 17 13    0    1 0.00000000 0.047602055 1.1635755 1.2111775   3.482006   1
#> 18 13    1    2 0.04760205 0.787029376 1.2111775 1.9506049   3.482006   1
#> 19 14    0    2 0.00000000 0.039312840 1.1645494 1.2038623  43.969894   1
#> 20 15    0    2 0.00000000 0.510728090 0.4173794 0.9281075  28.104000   1
#> 21 16    0    2 0.00000000 0.005959027 0.5619078 0.5678668  15.494744   1
#> 22 17    0    1 0.00000000 0.018787496 0.5491695 0.5679570  32.818580   1
#> 23 17    1    2 0.01878750 0.095444457 0.5679570 0.6446140  32.818580   1
#> 24 18    0    1 0.00000000 0.509938664 1.1668983 1.6768370  20.058654   1
#> 25 18    1    2 0.50993866 1.141412889 1.6768370 2.3083112  20.058654   1
#> 26 19    0    1 0.00000000 0.163696385 0.9949180 1.1586144   7.377779   1
#> 27 19    1    2 0.16369639 0.394728991 1.1586144 1.3896470   7.377779   1
#> 28 20    0    2 0.00000000 0.526957370 2.3844081 2.9113655   5.729975   1
#> 29 21    0    2 0.00000000 0.354918520 1.1883234 1.5432420  35.354131   2
#> 30 22    0    2 0.00000000 0.268112613 2.2186025 2.4867151   4.057648   2
#> 31 23    0    2 0.00000000 0.173935870 2.3080080 2.4819439  20.490825   2
#> 32 24    0    2 0.00000000 0.784650677 2.7637038 3.5483544  22.878374   2
#> 33 25    0    2 0.00000000 1.653583991 0.6168664 2.2704504  14.852909   2
#> 34 26    0    1 0.00000000 0.469706678 2.4771683 2.9468750  11.283897   2
#> 35 26    1    2 0.46970668 0.786725182 2.9468750 3.2638935  11.283897   2
#> 36 27    0    2 0.00000000 0.115660109 1.1911944 1.3068545   3.530013   2
#> 37 28    0    1 0.00000000 0.743637459 1.4657963 2.2094338  22.887217   2
#> 38 28    1    2 0.74363746 1.575873842 2.2094338 3.0416702  22.887217   2
#> 39 29    0    1 0.00000000 0.560799566 2.2141066 2.7749062  22.812513   2
#> 40 29    1    2 0.56079957 1.440325376 2.7749062 3.6544320  22.812513   2
#> 41 30    0    1 0.00000000 1.628128117 0.3786952 2.0068233  65.305047   2
#> 42 30    1    2 1.62812812 2.129508685 2.0068233 2.5082039  65.305047   2
#> 43 31    0    1 0.00000000 0.133929814 1.1768790 1.3108088   4.442713   2
#> 44 31    1    2 0.13392981 0.256939162 1.3108088 1.4338181   4.442713   2
#> 45 32    0 cens 0.00000000 0.410515591 1.8915645 2.3020801   2.302080   2
#> 46 33    0    1 0.00000000 0.290962737 2.5934850 2.8844478   8.395399   2
#> 47 33    1    2 0.29096274 1.126206605 2.8844478 3.7196917   8.395399   2
#> 48 34    0    2 0.00000000 0.158485849 0.9842618 1.1427477   1.524137   2
#> 49 35    0    1 0.00000000 0.218559636 0.3006945 0.5192541   8.732320   2
#> 50 35    1    2 0.21855964 0.275773453 0.5192541 0.5764679   8.732320   2
#> 51 36    0    2 0.00000000 0.619290848 2.6638455 3.2831363  27.433460   2
#> 52 37    0    2 0.00000000 0.046196359 0.5712379 0.6174343   7.803766   2
#> 53 38    0    2 0.00000000 0.237858357 0.1543380 0.3921964   5.995148   2
#> 54 39    0    2 0.00000000 0.466198020 1.2166859 1.6828840  10.501363   2
#> 55 40    0    1 0.00000000 0.152144411 0.9679881 1.1201325 107.536769   2
#> 56 40    1    2 0.15214441 0.724973832 1.1201325 1.6929619 107.536769   2
#> 
#> [[3]]
#>    id from to        entry         exit  entryAct   exitAct   censAct trt
#> 1   1    0  2 0.0000000000 0.0713746063 0.2522126 0.3235872 14.025087   1
#> 2   2    0  2 0.0000000000 0.0792380879 1.8778922 1.9571303  7.538700   1
#> 3   3    0  1 0.0000000000 0.2928811794 0.9488027 1.2416839 10.778614   1
#> 4   3    1  2 0.2928811794 0.9942501388 1.2416839 1.9430528 10.778614   1
#> 5   4    0  1 0.0000000000 0.1849724382 2.3869115 2.5718840 17.653398   1
#> 6   4    1  2 0.1849724382 1.3537912183 2.5718840 3.7407027 17.653398   1
#> 7   5    0  1 0.0000000000 0.3634652036 0.7079238 1.0713890  6.214253   1
#> 8   5    1  2 0.3634652036 1.2209592176 1.0713890 1.9288830  6.214253   1
#> 9   6    0  2 0.0000000000 0.3144854217 1.0388312 1.3533166 22.061282   1
#> 10  7    0  1 0.0000000000 0.0561270364 2.6418532 2.6979803  9.722939   1
#> 11  7    1  2 0.0561270364 1.1620006549 2.6979803 3.8038539  9.722939   1
#> 12  8    0  1 0.0000000000 0.1264685247 1.6086229 1.7350914 12.503362   1
#> 13  8    1  2 0.1264685247 0.2120006042 1.7350914 1.8206235 12.503362   1
#> 14  9    0  1 0.0000000000 0.0002267938 2.3053296 2.3055564  8.326828   1
#> 15  9    1  2 0.0002267938 0.9879803233 2.3055564 3.2933099  8.326828   1
#> 16 10    0  2 0.0000000000 1.1515586052 0.1112250 1.2627836  6.774442   1
#> 17 11    0  1 0.0000000000 0.1888306261 0.1137870 0.3026176  4.033839   1
#> 18 11    1  2 0.1888306261 0.4633866458 0.3026176 0.5771736  4.033839   1
#> 19 12    0  2 0.0000000000 1.3979969773 1.8072656 3.2052626 51.114306   1
#> 20 13    0  1 0.0000000000 0.2655471381 0.6831051 0.9486522 13.008625   1
#> 21 13    1  2 0.2655471381 2.1143492608 0.9486522 2.7974543 13.008625   1
#> 22 14    0  2 0.0000000000 0.2331615993 0.7215000 0.9546616 10.276832   1
#> 23 15    0  1 0.0000000000 0.4765772420 1.7846414 2.2612187  4.747279   1
#> 24 15    1  2 0.4765772420 2.1400280216 2.2612187 3.9246695  4.747279   1
#> 25 16    0  1 0.0000000000 0.0567544109 1.1591137 1.2158681  4.848823   1
#> 26 16    1  2 0.0567544109 0.6177408357 1.2158681 1.7768546  4.848823   1
#> 27 17    0  1 0.0000000000 0.2746114073 2.2813516 2.5559630  4.842559   1
#> 28 17    1  2 0.2746114073 0.3477922345 2.5559630 2.6291439  4.842559   1
#> 29 18    0  1 0.0000000000 0.3044215191 0.3356084 0.6400299 58.498594   1
#> 30 18    1  2 0.3044215191 1.2617436682 0.6400299 1.5973521 58.498594   1
#> 31 19    0  1 0.0000000000 0.3999327490 2.8414208 3.2413536 23.346637   1
#> 32 19    1  2 0.3999327490 0.9754693118 3.2413536 3.8168901 23.346637   1
#> 33 20    0  2 0.0000000000 0.1724586275 1.8567352 2.0291939  6.557996   1
#> 34 21    0  2 0.0000000000 0.2242285109 2.6555759 2.8798044  8.043555   2
#> 35 22    0  1 0.0000000000 0.2127519456 0.1425367 0.3552886  8.711440   2
#> 36 22    1  2 0.2127519456 1.9382209300 0.3552886 2.0807576  8.711440   2
#> 37 23    0  1 0.0000000000 0.0270080114 1.9181645 1.9451725  2.412370   2
#> 38 23    1  2 0.0270080114 0.3069435955 1.9451725 2.2251081  2.412370   2
#> 39 24    0  1 0.0000000000 0.2370481570 2.3586055 2.5956537 23.383883   2
#> 40 24    1  2 0.2370481570 0.4703904213 2.5956537 2.8289959 23.383883   2
#> 41 25    0  1 0.0000000000 0.3482386995 0.9617543 1.3099930  4.815768   2
#> 42 25    1  2 0.3482386995 0.9912434309 1.3099930 1.9529978  4.815768   2
#> 43 26    0  2 0.0000000000 0.7361874439 2.6614874 3.3976748  5.251791   2
#> 44 27    0  2 0.0000000000 0.7908285443 1.3829883 2.1738169 23.326143   2
#> 45 28    0  2 0.0000000000 0.0052921047 2.7029914 2.7082835  5.384365   2
#> 46 29    0  2 0.0000000000 0.3696802097 0.8893163 1.2589965 10.952388   2
#> 47 30    0  2 0.0000000000 0.1560964280 1.7115768 1.8676732 31.264657   2
#> 48 31    0  2 0.0000000000 0.0171506912 2.5434955 2.5606462 14.249264   2
#> 49 32    0  2 0.0000000000 0.1929068099 1.7353244 1.9282312  2.213021   2
#> 50 33    0  1 0.0000000000 0.1353165640 1.2587383 1.3940549 12.698245   2
#> 51 33    1  2 0.1353165640 1.0188601197 1.3940549 2.2775985 12.698245   2
#> 52 34    0  2 0.0000000000 1.5872052891 1.6001704 3.1873756 29.867539   2
#> 53 35    0  2 0.0000000000 1.3125507615 2.6840778 3.9966286  5.584007   2
#> 54 36    0  2 0.0000000000 0.4162046035 2.3843189 2.8005235  9.732216   2
#> 55 37    0  2 0.0000000000 1.4555483033 0.8205429 2.2760912 36.393381   2
#> 56 38    0  1 0.0000000000 0.9147900845 2.4101717 3.3249617 57.913621   2
#> 57 38    1  2 0.9147900845 1.3041337560 3.3249617 3.7143054 57.913621   2
#> 58 39    0  2 0.0000000000 0.0555747869 0.7393658 0.7949406 34.937113   2
#> 59 40    0  2 0.0000000000 0.3943386436 0.9695367 1.3638753  6.705119   2
#> 
#> [[4]]
#>    id from   to      entry         exit    entryAct   exitAct   censAct trt
#> 1   1    0    2 0.00000000 0.2410986442 0.981702251 1.2228009  2.952092   1
#> 2   2    0    1 0.00000000 0.0432759703 0.871870763 0.9151467 33.580123   1
#> 3   2    1    2 0.04327597 1.1319203262 0.915146733 2.0037911 33.580123   1
#> 4   3    0 cens 0.00000000 0.3340910602 2.819007861 3.1530989  3.153099   1
#> 5   4    0    2 0.00000000 0.1286766675 2.640844855 2.7695215  6.324510   1
#> 6   5    0    2 0.00000000 0.6569196895 0.034888646 0.6918083 54.178804   1
#> 7   6    0    2 0.00000000 0.1472025899 0.283570588 0.4307732 10.834122   1
#> 8   7    0    2 0.00000000 0.7662725414 1.546054828 2.3123274 33.788079   1
#> 9   8    0    1 0.00000000 0.0103786084 1.773422543 1.7838012 15.817919   1
#> 10  8    1    2 0.01037861 0.0769419978 1.783801152 1.8503645 15.817919   1
#> 11  9    0    2 0.00000000 0.2876315947 2.267463900 2.5550955  2.667143   1
#> 12 10    0    1 0.00000000 0.2414408720 1.375794810 1.6172357  8.756231   1
#> 13 10    1    2 0.24144087 2.4790636514 1.617235682 3.8548585  8.756231   1
#> 14 11    0    1 0.00000000 0.7788710668 2.306212885 3.0850840 13.233571   1
#> 15 11    1    2 0.77887107 0.7822318384 3.085083951 3.0884447 13.233571   1
#> 16 12    0    1 0.00000000 0.2470910589 2.822382450 3.0694735 60.830550   1
#> 17 12    1    2 0.24709106 0.6777033818 3.069473509 3.5000858 60.830550   1
#> 18 13    0    1 0.00000000 0.0931432113 2.650214293 2.7433575 18.249118   1
#> 19 13    1    2 0.09314321 2.7461783458 2.743357505 5.3963926 18.249118   1
#> 20 14    0    2 0.00000000 0.2635578950 0.775950156 1.0395081  1.192922   1
#> 21 15    0    2 0.00000000 0.0035450488 0.282266245 0.2858113  7.565936   1
#> 22 16    0    2 0.00000000 0.0343161680 0.187464524 0.2217807 14.587169   1
#> 23 17    0    2 0.00000000 0.1420633773 0.535214855 0.6772782 14.072931   1
#> 24 18    0    2 0.00000000 0.2919005722 1.691360939 1.9832615  4.680850   1
#> 25 19    0    1 0.00000000 0.1336806774 0.002465959 0.1361466  4.985135   1
#> 26 19    1    2 0.13368068 0.4954993259 0.136146637 0.4979653  4.985135   1
#> 27 20    0    2 0.00000000 0.4987507578 1.383390286 1.8821410  3.864279   1
#> 28 21    0    2 0.00000000 0.3108055841 1.108384993 1.4191906  6.488981   2
#> 29 22    0    2 0.00000000 0.0296386482 1.115127469 1.1447661 35.599421   2
#> 30 23    0    1 0.00000000 0.1103444532 0.684055658 0.7944001  6.379908   2
#> 31 23    1    2 0.11034445 0.3885954714 0.794400111 1.0726511  6.379908   2
#> 32 24    0    2 0.00000000 0.3786637830 2.255959040 2.6346228  4.078012   2
#> 33 25    0    1 0.00000000 0.1955502860 0.045346345 0.2408966  8.498013   2
#> 34 25    1    2 0.19555029 0.4072976185 0.240896631 0.4526440  8.498013   2
#> 35 26    0    2 0.00000000 1.2602046573 1.139314443 2.3995191 22.935091   2
#> 36 27    0    1 0.00000000 0.0253007462 1.109335964 1.1346367 12.731233   2
#> 37 27    1    2 0.02530075 0.9876996308 1.134636710 2.0970356 12.731233   2
#> 38 28    0    2 0.00000000 0.0001486396 1.307943312 1.3080920 29.748809   2
#> 39 29    0    2 0.00000000 0.1488971253 0.847559424 0.9964565  4.831768   2
#> 40 30    0    2 0.00000000 0.1065229593 2.294964366 2.4014873 51.405341   2
#> 41 31    0    1 0.00000000 0.5072148269 2.489700025 2.9969149  7.602202   2
#> 42 31    1    2 0.50721483 0.8808151622 2.996914852 3.3705152  7.602202   2
#> 43 32    0    2 0.00000000 0.8702256858 2.698787467 3.5690132 87.131527   2
#> 44 33    0    2 0.00000000 0.0131704040 0.779711168 0.7928816  3.457760   2
#> 45 34    0    1 0.00000000 0.5256586610 1.929826306 2.4554850 16.060009   2
#> 46 34    1    2 0.52565866 0.8682492314 2.455484967 2.7980755 16.060009   2
#> 47 35    0    2 0.00000000 0.0793097535 2.032602950 2.1119127  5.859646   2
#> 48 36    0    2 0.00000000 0.6684316624 0.576163331 1.2445950  7.454743   2
#> 49 37    0    2 0.00000000 0.0685444380 1.590595764 1.6591402 17.087538   2
#> 50 38    0    1 0.00000000 0.4902738956 1.781466700 2.2717406 11.826361   2
#> 51 38    1    2 0.49027390 0.5390640410 2.271740596 2.3205307 11.826361   2
#> 52 39    0    1 0.00000000 0.3193638270 0.308769981 0.6281338 38.968055   2
#> 53 39    1    2 0.31936383 0.5944596499 0.628133808 0.9032296 38.968055   2
#> 54 40    0    2 0.00000000 0.5307724009 1.670971908 2.2017443  5.961288   2
#> 
#> [[5]]
#>    id from   to      entry         exit   entryAct   exitAct   censAct trt
#> 1   1    0    1 0.00000000 0.0332999488 1.68885656 1.7221565  6.738915   1
#> 2   1    1    2 0.03329995 2.5591184818 1.72215651 4.2479750  6.738915   1
#> 3   2    0    1 0.00000000 0.2636246081 0.73782300 1.0014476 33.543444   1
#> 4   2    1    2 0.26362461 2.3384513673 1.00144761 3.0762744 33.543444   1
#> 5   3    0    1 0.00000000 0.3192309551 2.16401042 2.4832414 20.955361   1
#> 6   3    1    2 0.31923096 1.3176604081 2.48324138 3.4816708 20.955361   1
#> 7   4    0    1 0.00000000 0.4301024837 1.57584591 2.0059484 10.177052   1
#> 8   4    1    2 0.43010248 0.9177964123 2.00594840 2.4936423 10.177052   1
#> 9   5    0    2 0.00000000 0.1073236999 0.27892951 0.3862532  1.849727   1
#> 10  6    0    1 0.00000000 0.2184061612 0.43001827 0.6484244  2.548339   1
#> 11  6    1    2 0.21840616 0.5986522948 0.64842443 1.0286706  2.548339   1
#> 12  7    0    2 0.00000000 0.2885957510 0.41653084 0.7051266  5.432299   1
#> 13  8    0    1 0.00000000 0.3110182152 2.83019489 3.1412131 50.580754   1
#> 14  8    1    2 0.31101822 0.4591634038 3.14121310 3.2893583 50.580754   1
#> 15  9    0    2 0.00000000 0.9871455450 0.07454948 1.0616950 51.929095   1
#> 16 10    0    2 0.00000000 0.0700754248 1.73352670 1.8036021 11.113885   1
#> 17 11    0    2 0.00000000 0.0148296548 0.29177314 0.3066028  1.619974   1
#> 18 12    0    1 0.00000000 0.0385860797 1.45550172 1.4940878 63.963362   1
#> 19 12    1    2 0.03858608 0.2024578992 1.49408780 1.6579596 63.963362   1
#> 20 13    0    2 0.00000000 0.0057884263 0.53548956 0.5412780  3.274291   1
#> 21 14    0    2 0.00000000 0.9373104972 1.44920540 2.3865159 14.955906   1
#> 22 15    0    2 0.00000000 0.5082140903 1.10050147 1.6087156  2.192451   1
#> 23 16    0    2 0.00000000 0.3444093398 1.68190093 2.0263103 12.540869   1
#> 24 17    0    2 0.00000000 0.1756765585 2.63069727 2.8063738 18.163925   1
#> 25 18    0    2 0.00000000 0.1547321886 1.66730836 1.8220405 10.881232   1
#> 26 19    0    2 0.00000000 0.1469674060 2.32257739 2.4695448 35.240483   1
#> 27 20    0    2 0.00000000 0.3657595210 2.19295156 2.5587111 48.957341   1
#> 28 21    0    1 0.00000000 0.0842679335 0.87288477 0.9571527  4.597237   2
#> 29 21    1    2 0.08426793 0.3594287113 0.95715270 1.2323135  4.597237   2
#> 30 22    0    2 0.00000000 0.2866589560 2.65364068 2.9402996  3.823342   2
#> 31 23    0    1 0.00000000 0.4394935112 0.74593618 1.1854297 15.202827   2
#> 32 23    1    2 0.43949351 0.5647225047 1.18542969 1.3106587 15.202827   2
#> 33 24    0    2 0.00000000 1.8874770545 1.26606557 3.1535426 41.674347   2
#> 34 25    0    1 0.00000000 0.0455841890 2.02184233 2.0674265  5.707270   2
#> 35 25    1    2 0.04558419 0.9429468934 2.06742652 2.9647892  5.707270   2
#> 36 26    0    1 0.00000000 0.5144061244 2.57984162 3.0942477  3.131123   2
#> 37 26    1    2 0.51440612 0.5164233540 3.09424774 3.0962650  3.131123   2
#> 38 27    0    2 0.00000000 0.0005693736 1.46606603 1.4666354 14.926529   2
#> 39 28    0    1 0.00000000 0.6341506733 1.43006866 2.0642193 30.952348   2
#> 40 28    1    2 0.63415067 2.2070719245 2.06421934 3.6371406 30.952348   2
#> 41 29    0    1 0.00000000 0.1676662615 0.14226202 0.3099283 10.298686   2
#> 42 29    1    2 0.16766626 0.8555265340 0.30992828 0.9977886 10.298686   2
#> 43 30    0    2 0.00000000 1.3784311037 0.71254331 2.0909744 56.762391   2
#> 44 31    0    1 0.00000000 0.0945247874 2.25280301 2.3473278 39.718569   2
#> 45 31    1    2 0.09452479 0.2659120726 2.34732780 2.5187151 39.718569   2
#> 46 32    0    2 0.00000000 0.7903920482 0.56618355 1.3565756  6.197933   2
#> 47 33    0    2 0.00000000 0.2906037904 2.51740811 2.8080119 30.479469   2
#> 48 34    0 cens 0.00000000 0.0380374622 1.04090797 1.0789454  1.078945   2
#> 49 35    0    1 0.00000000 0.1882095953 0.77128564 0.9594952  9.422382   2
#> 50 35    1    2 0.18820960 0.8322813046 0.95949524 1.6035669  9.422382   2
#> 51 36    0    2 0.00000000 0.6224154408 1.68564151 2.3080569 32.703041   2
#> 52 37    0    2 0.00000000 0.2146726244 0.47271595 0.6873886 10.841356   2
#> 53 38    0    1 0.00000000 0.2651092776 1.39734224 1.6624515  2.316457   2
#> 54 38    1    2 0.26510928 0.6461735571 1.66245152 2.0435158  2.316457   2
#> 55 39    0    1 0.00000000 0.0493714249 2.75252470 2.8018961  3.225759   2
#> 56 39    1 cens 0.04937142 0.4732346311 2.80189613 3.2257593  3.225759   2
#> 57 40    0    2 0.00000000 1.4296042193 1.65600838 3.0856126 26.743329   2
#> 
#> [[6]]
#>    id from to      entry       exit   entryAct   exitAct   censAct trt
#> 1   1    0  2 0.00000000 0.31910775 0.10926129 0.4283690 10.178046   1
#> 2   2    0  1 0.00000000 0.01456238 2.44170945 2.4562718  5.171734   1
#> 3   2    1  2 0.01456238 0.61010779 2.45627183 3.0518172  5.171734   1
#> 4   3    0  1 0.00000000 0.47985426 1.75172996 2.2315842  3.201626   1
#> 5   3    1  2 0.47985426 0.61941713 2.23158423 2.3711471  3.201626   1
#> 6   4    0  2 0.00000000 0.83961224 1.30566490 2.1452771 10.397126   1
#> 7   5    0  1 0.00000000 0.49698387 0.74485444 1.2418383 27.134699   1
#> 8   5    1  2 0.49698387 1.99454196 1.24183831 2.7393964 27.134699   1
#> 9   6    0  2 0.00000000 0.19243376 0.72422237 0.9166561 22.582336   1
#> 10  7    0  2 0.00000000 0.03179525 0.50714049 0.5389357 11.022942   1
#> 11  8    0  1 0.00000000 0.20416365 1.84802987 2.0521935  2.309456   1
#> 12  8    1  2 0.20416365 0.43953900 2.05219352 2.2875689  2.309456   1
#> 13  9    0  2 0.00000000 0.05128322 0.62200592 0.6732891 29.896489   1
#> 14 10    0  1 0.00000000 0.01149837 0.07194894 0.0834473 37.661026   1
#> 15 10    1  2 0.01149837 0.13835916 0.08344730 0.2103081 37.661026   1
#> 16 11    0  2 0.00000000 0.05026240 1.91903046 1.9692929 27.804553   1
#> 17 12    0  2 0.00000000 0.16370250 0.22680711 0.3905096 18.494981   1
#> 18 13    0  2 0.00000000 0.10434070 0.04368865 0.1480293 17.800040   1
#> 19 14    0  2 0.00000000 0.26904635 0.63565901 0.9047054  5.308293   1
#> 20 15    0  1 0.00000000 0.31049670 0.95780612 1.2683028 23.740501   1
#> 21 15    1  2 0.31049670 1.89948223 1.26830282 2.8572883 23.740501   1
#> 22 16    0  2 0.00000000 0.21911925 0.03557061 0.2546899  5.852499   1
#> 23 17    0  1 0.00000000 0.47667126 0.44388214 0.9205534  3.531466   1
#> 24 17    1  2 0.47667126 0.80841498 0.92055339 1.2522971  3.531466   1
#> 25 18    0  1 0.00000000 0.15177837 0.36651944 0.5182978 12.593604   1
#> 26 18    1  2 0.15177837 0.74597451 0.51829781 1.1124939 12.593604   1
#> 27 19    0  2 0.00000000 0.08614689 0.47998545 0.5661323 25.150757   1
#> 28 20    0  2 0.00000000 1.76177635 2.62927278 4.3910491 39.016045   1
#> 29 21    0  2 0.00000000 0.15946480 0.34028502 0.4997498  6.658468   2
#> 30 22    0  2 0.00000000 0.44978088 2.44249527 2.8922762 84.532024   2
#> 31 23    0  1 0.00000000 0.20918819 0.14674985 0.3559380  4.907253   2
#> 32 23    1  2 0.20918819 0.42430431 0.35593804 0.5710542  4.907253   2
#> 33 24    0  2 0.00000000 0.39083416 1.81531432 2.2061485 13.204293   2
#> 34 25    0  1 0.00000000 0.68843072 1.77008588 2.4585166 19.484967   2
#> 35 25    1  2 0.68843072 0.81487062 2.45851660 2.5849565 19.484967   2
#> 36 26    0  1 0.00000000 0.15525099 0.52802770 0.6832787 12.627945   2
#> 37 26    1  2 0.15525099 1.79260886 0.68327869 2.3206366 12.627945   2
#> 38 27    0  2 0.00000000 0.09586063 0.16295599 0.2588166 66.695923   2
#> 39 28    0  1 0.00000000 0.30831955 1.43571270 1.7440322 35.945748   2
#> 40 28    1  2 0.30831955 1.29014098 1.74403224 2.7258537 35.945748   2
#> 41 29    0  1 0.00000000 0.08108648 0.95544977 1.0365362 23.950632   2
#> 42 29    1  2 0.08108648 0.57561808 1.03653624 1.5310678 23.950632   2
#> 43 30    0  2 0.00000000 0.13831762 2.07038943 2.2087071 29.749732   2
#> 44 31    0  2 0.00000000 0.42418270 2.18038347 2.6045662 21.626253   2
#> 45 32    0  1 0.00000000 0.18123904 2.73575771 2.9169967 23.363804   2
#> 46 32    1  2 0.18123904 1.41826828 2.91699675 4.1540260 23.363804   2
#> 47 33    0  2 0.00000000 0.56909452 0.70177305 1.2708676  2.868643   2
#> 48 34    0  2 0.00000000 0.11168271 0.53987469 0.6515574 16.575232   2
#> 49 35    0  1 0.00000000 0.02020601 2.34231943 2.3625254 11.972886   2
#> 50 35    1  2 0.02020601 0.15266445 2.36252544 2.4949839 11.972886   2
#> 51 36    0  1 0.00000000 1.53043801 1.00569362 2.5361316  4.668653   2
#> 52 36    1  2 1.53043801 1.96501847 2.53613163 2.9707121  4.668653   2
#> 53 37    0  2 0.00000000 0.66566040 1.20887575 1.8745361 12.396305   2
#> 54 38    0  2 0.00000000 0.85031411 2.40114758 3.2514617  6.295855   2
#> 55 39    0  2 0.00000000 0.30552495 1.37017648 1.6757014  2.007219   2
#> 56 40    0  1 0.00000000 0.58891406 0.92306765 1.5119817  4.265156   2
#> 57 40    1  2 0.58891406 0.76281222 1.51198170 1.6858799  4.265156   2
#> 
#> [[7]]
#>    id from   to      entry       exit  entryAct   exitAct    censAct trt
#> 1   1    0    2 0.00000000 0.23713597 1.8252590 2.0623950  7.2744124   1
#> 2   2    0    1 0.00000000 0.03361286 0.3220804 0.3556932  0.9047256   1
#> 3   2    1    2 0.03361286 0.05817040 0.3556932 0.3802508  0.9047256   1
#> 4   3    0    1 0.00000000 0.04966008 1.0339738 1.0836339 20.6272727   1
#> 5   3    1    2 0.04966008 0.36676278 1.0836339 1.4007366 20.6272727   1
#> 6   4    0    2 0.00000000 0.09040027 2.4021185 2.4925187 42.6750138   1
#> 7   5    0    1 0.00000000 0.22420981 2.2287212 2.4529310  4.9269626   1
#> 8   5    1    2 0.22420981 0.39352561 2.4529310 2.6222468  4.9269626   1
#> 9   6    0    2 0.00000000 0.80772762 2.0205934 2.8283211 40.5142418   1
#> 10  7    0    2 0.00000000 0.28840385 2.1401341 2.4285380  6.3574253   1
#> 11  8    0    1 0.00000000 0.96632146 0.5375291 1.5038506  1.5094680   1
#> 12  8    1 cens 0.96632146 0.97193893 1.5038506 1.5094680  1.5094680   1
#> 13  9    0    2 0.00000000 0.10233572 2.3965451 2.4988808  6.6445660   1
#> 14 10    0    1 0.00000000 0.44510154 0.7729241 1.2180256  1.5556914   1
#> 15 10    1    2 0.44510154 0.52567944 1.2180256 1.2986035  1.5556914   1
#> 16 11    0    1 0.00000000 0.26910671 1.0479737 1.3170804  6.7657573   1
#> 17 11    1    2 0.26910671 0.50561139 1.3170804 1.5535851  6.7657573   1
#> 18 12    0    2 0.00000000 0.38798333 0.9737994 1.3617828 21.8140785   1
#> 19 13    0    2 0.00000000 0.11064259 0.7776597 0.8883023  4.7940685   1
#> 20 14    0    1 0.00000000 0.64966795 1.6443023 2.2939703 26.6493531   1
#> 21 14    1    2 0.64966795 0.72907323 2.2939703 2.3733756 26.6493531   1
#> 22 15    0    1 0.00000000 0.11141290 2.5282248 2.6396377 14.5591124   1
#> 23 15    1    2 0.11141290 0.41776148 2.6396377 2.9459863 14.5591124   1
#> 24 16    0    1 0.00000000 0.04579814 1.0394231 1.0852212 86.8895340   1
#> 25 16    1    2 0.04579814 1.89557262 1.0852212 2.9349957 86.8895340   1
#> 26 17    0    2 0.00000000 0.20111616 1.9874431 2.1885592  3.6354171   1
#> 27 18    0    2 0.00000000 0.86074021 2.2738236 3.1345638 38.3980915   1
#> 28 19    0    2 0.00000000 0.43166341 1.9995116 2.4311750 17.7034800   1
#> 29 20    0    2 0.00000000 0.10260895 1.0996593 1.2022682 14.5118318   1
#> 30 21    0    1 0.00000000 0.67923401 2.6836286 3.3628626  6.7755020   2
#> 31 21    1    2 0.67923401 1.07105513 3.3628626 3.7546837  6.7755020   2
#> 32 22    0    2 0.00000000 0.05203213 2.3159447 2.3679768 79.5504088   2
#> 33 23    0    2 0.00000000 0.75551079 0.8507753 1.6062861 49.2751438   2
#> 34 24    0    2 0.00000000 0.63510684 1.8113438 2.4464506 39.6676240   2
#> 35 25    0    1 0.00000000 0.48921432 0.9809311 1.4701454 26.5694130   2
#> 36 25    1    2 0.48921432 2.45155438 1.4701454 3.4324855 26.5694130   2
#> 37 26    0    1 0.00000000 0.39119775 1.5670086 1.9582064  7.9505087   2
#> 38 26    1    2 0.39119775 2.15036802 1.9582064 3.7173767  7.9505087   2
#> 39 27    0    1 0.00000000 1.11466023 0.7542657 1.8689259 14.6302708   2
#> 40 27    1    2 1.11466023 1.13410502 1.8689259 1.8883707 14.6302708   2
#> 41 28    0    1 0.00000000 0.70358296 0.7639384 1.4675213  9.4135827   2
#> 42 28    1    2 0.70358296 0.71726813 1.4675213 1.4812065  9.4135827   2
#> 43 29    0    2 0.00000000 0.23476457 1.9724856 2.2072501 69.1746629   2
#> 44 30    0    1 0.00000000 0.74149200 1.6560837 2.3975757  4.3833475   2
#> 45 30    1    2 0.74149200 0.98027502 2.3975757 2.6363588  4.3833475   2
#> 46 31    0    2 0.00000000 0.05022066 1.8977037 1.9479244 31.5020746   2
#> 47 32    0    2 0.00000000 0.21947915 1.1621516 1.3816307  5.0432077   2
#> 48 33    0    2 0.00000000 0.19349331 1.2919865 1.4854799 22.3093186   2
#> 49 34    0    2 0.00000000 0.01571694 1.5578480 1.5735649  9.0207176   2
#> 50 35    0    2 0.00000000 0.15020660 0.3271609 0.4773675 53.3337255   2
#> 51 36    0    1 0.00000000 0.13716677 0.8047313 0.9418981  2.8809992   2
#> 52 36    1    2 0.13716677 0.30948268 0.9418981 1.1142140  2.8809992   2
#> 53 37    0    2 0.00000000 0.43057446 0.7329241 1.1634986 15.6320572   2
#> 54 38    0    2 0.00000000 0.50425350 2.5219844 3.0262379 36.3868541   2
#> 55 39    0    2 0.00000000 0.08381527 2.3990472 2.4828625 23.4604056   2
#> 56 40    0    2 0.00000000 2.51112154 2.4440377 4.9551593 24.3227898   2
#> 
#> [[8]]
#>    id from   to      entry        exit   entryAct   exitAct   censAct trt
#> 1   1    0    2 0.00000000 0.310151741 2.84740783 3.1575596  6.799396   1
#> 2   2    0    2 0.00000000 0.290583550 1.05786567 1.3484492 20.449987   1
#> 3   3    0    2 0.00000000 0.179078147 0.42956313 0.6086413  1.776445   1
#> 4   4    0    2 0.00000000 0.607344411 1.54870156 2.1560460 22.581290   1
#> 5   5    0    2 0.00000000 0.690308459 1.22065822 1.9109667 29.577431   1
#> 6   6    0    1 0.00000000 0.228614953 0.66040333 0.8890183  5.246746   1
#> 7   6    1    2 0.22861495 0.400316945 0.88901828 1.0607203  5.246746   1
#> 8   7    0    2 0.00000000 0.374055349 2.45552026 2.8295756 30.105786   1
#> 9   8    0    2 0.00000000 0.288647331 0.03679664 0.3254440  1.383316   1
#> 10  9    0    2 0.00000000 0.642159267 1.81051897 2.4526782 48.063579   1
#> 11 10    0    2 0.00000000 0.806135471 1.06170914 1.8678446  6.202371   1
#> 12 11    0    1 0.00000000 0.532134809 1.21627317 1.7484080  6.845063   1
#> 13 11    1    2 0.53213481 1.180460735 1.74840798 2.3967339  6.845063   1
#> 14 12    0    1 0.00000000 0.285471354 1.95139218 2.2368635 10.648473   1
#> 15 12    1    2 0.28547135 0.453937538 2.23686353 2.4053297 10.648473   1
#> 16 13    0    2 0.00000000 0.751004535 0.27517552 1.0261801 10.811445   1
#> 17 14    0    2 0.00000000 0.094866684 2.15611031 2.2509770 24.160941   1
#> 18 15    0    2 0.00000000 0.371633829 1.11501013 1.4866440  5.075530   1
#> 19 16    0    1 0.00000000 0.392599439 1.64107276 2.0336722  3.763091   1
#> 20 16    1    2 0.39259944 1.571951523 2.03367220 3.2130243  3.763091   1
#> 21 17    0    2 0.00000000 0.400110984 1.21917438 1.6192854 41.195136   1
#> 22 18    0    2 0.00000000 0.054794669 0.09584648 0.1506412 10.896583   1
#> 23 19    0    1 0.00000000 0.488007314 1.33546237 1.8234697  1.919626   1
#> 24 19    1 cens 0.48800731 0.584163149 1.82346968 1.9196255  1.919626   1
#> 25 20    0    2 0.00000000 0.027827071 0.69781666 0.7256437 42.566845   1
#> 26 21    0    1 0.00000000 0.048617845 2.46607093 2.5146888  3.072793   2
#> 27 21    1    2 0.04861784 0.245911354 2.51468878 2.7119823  3.072793   2
#> 28 22    0    1 0.00000000 0.515465728 2.13648268 2.6519484 94.831941   2
#> 29 22    1    2 0.51546573 0.915497643 2.65194841 3.0519803 94.831941   2
#> 30 23    0    2 0.00000000 0.045939544 2.68641602 2.7323556 51.506097   2
#> 31 24    0    2 0.00000000 1.280302943 2.62921805 3.9095210 44.704811   2
#> 32 25    0    1 0.00000000 0.072163515 0.45600299 0.5281665 16.113644   2
#> 33 25    1    2 0.07216352 0.186729417 0.52816651 0.6427324 16.113644   2
#> 34 26    0 cens 0.00000000 0.045812317 2.50435107 2.5501634  2.550163   2
#> 35 27    0    1 0.00000000 0.037493647 2.04327457 2.0807682 10.265634   2
#> 36 27    1    2 0.03749365 1.215557261 2.08076822 3.2588318 10.265634   2
#> 37 28    0    1 0.00000000 0.466083809 1.27092170 1.7370055 49.292839   2
#> 38 28    1    2 0.46608381 0.546674699 1.73700550 1.8175964 49.292839   2
#> 39 29    0    2 0.00000000 0.284626623 0.94074121 1.2253678 15.704232   2
#> 40 30    0    2 0.00000000 0.007240986 0.84237828 0.8496193 54.110630   2
#> 41 31    0    2 0.00000000 1.456656721 1.33771461 2.7943713  9.215072   2
#> 42 32    0    2 0.00000000 0.082568450 0.40090744 0.4834759  5.165432   2
#> 43 33    0    1 0.00000000 0.268202326 2.72797024 2.9961726 31.217419   2
#> 44 33    1    2 0.26820233 0.343085594 2.99617257 3.0710558 31.217419   2
#> 45 34    0    1 0.00000000 0.357882220 1.40741485 1.7652971  7.693002   2
#> 46 34    1    2 0.35788222 1.365384767 1.76529707 2.7727996  7.693002   2
#> 47 35    0 cens 0.00000000 0.268598086 2.69017703 2.9587751  2.958775   2
#> 48 36    0    1 0.00000000 0.101104472 1.16579979 1.2669043 20.762987   2
#> 49 36    1    2 0.10110447 0.238665014 1.26690426 1.4044648 20.762987   2
#> 50 37    0    2 0.00000000 0.063105582 0.63017737 0.6932830  2.153837   2
#> 51 38    0    2 0.00000000 0.144081977 1.60599765 1.7500796 26.860895   2
#> 52 39    0    2 0.00000000 0.231100226 1.25628551 1.4873857 16.291715   2
#> 53 40    0    1 0.00000000 0.102080034 0.77976737 0.8818474  9.791589   2
#> 54 40    1    2 0.10208003 0.206078978 0.88184740 0.9858463  9.791589   2
#> 
#> [[9]]
#>    id from   to        entry         exit   entryAct   exitAct   censAct trt
#> 1   1    0    2 0.0000000000 0.4712528662 1.38305888 1.8543117  2.088719   1
#> 2   2    0    2 0.0000000000 0.0341122078 0.97511271 1.0092249  9.123580   1
#> 3   3    0    2 0.0000000000 0.4149434796 0.55369880 0.9686423 28.400391   1
#> 4   4    0    2 0.0000000000 0.6252475457 0.44336996 1.0686175  9.530231   1
#> 5   5    0    2 0.0000000000 0.1227371627 1.20022698 1.3229641 27.315522   1
#> 6   6    0    1 0.0000000000 0.0126862462 2.61196249 2.6246487  7.648715   1
#> 7   6    1    2 0.0126862462 1.5717055030 2.62464874 4.1836680  7.648715   1
#> 8   7    0    2 0.0000000000 0.0095227525 1.72300653 1.7325293 14.825508   1
#> 9   8    0    2 0.0000000000 0.2171286421 1.09755303 1.3146817  3.813328   1
#> 10  9    0    2 0.0000000000 0.2272809958 0.18417977 0.4114608 27.089509   1
#> 11 10    0    1 0.0000000000 0.4180034979 2.02245093 2.4404544 14.643702   1
#> 12 10    1    2 0.4180034979 0.8225176567 2.44045443 2.8449686 14.643702   1
#> 13 11    0    1 0.0000000000 0.0165026856 2.18116616 2.1976688  2.521556   1
#> 14 11    1 cens 0.0165026856 0.3403900350 2.19766885 2.5215562  2.521556   1
#> 15 12    0    2 0.0000000000 0.0246163610 1.81814737 1.8427637  3.607772   1
#> 16 13    0    2 0.0000000000 0.3864389662 1.70708903 2.0935280 17.729165   1
#> 17 14    0    2 0.0000000000 0.9313199102 2.21380388 3.1451238  7.671797   1
#> 18 15    0    1 0.0000000000 0.3863041907 2.52990502 2.9162092 15.249421   1
#> 19 15    1    2 0.3863041907 0.4025894051 2.91620921 2.9324944 15.249421   1
#> 20 16    0    2 0.0000000000 0.0771254657 2.06071420 2.1378397 24.035245   1
#> 21 17    0    2 0.0000000000 0.0608738581 2.39412308 2.4549969  7.491855   1
#> 22 18    0    1 0.0000000000 0.1354568749 0.62091639 0.7563733 42.477157   1
#> 23 18    1    2 0.1354568749 0.8619044766 0.75637326 1.4828209 42.477157   1
#> 24 19    0    1 0.0000000000 0.2571425530 0.84374007 1.1008826 48.242116   1
#> 25 19    1    2 0.2571425530 1.8901489506 1.10088262 2.7338890 48.242116   1
#> 26 20    0    1 0.0000000000 0.3081903494 2.74545841 3.0536488  3.250680   1
#> 27 20    1 cens 0.3081903494 0.5052217977 3.05364876 3.2506802  3.250680   1
#> 28 21    0    2 0.0000000000 0.2784890690 0.82290229 1.1013914 17.114072   2
#> 29 22    0    2 0.0000000000 0.3579176343 1.85399571 2.2119133 61.382354   2
#> 30 23    0    2 0.0000000000 1.1656533053 0.14321663 1.3088699 28.955028   2
#> 31 24    0    1 0.0000000000 0.1296996545 1.68012329 1.8098229 22.426566   2
#> 32 24    1    2 0.1296996545 0.1533290307 1.80982294 1.8334523 22.426566   2
#> 33 25    0    2 0.0000000000 0.6495941597 0.23753713 0.8871313 19.770969   2
#> 34 26    0    1 0.0000000000 0.0132166817 0.71495343 0.7281701  2.984517   2
#> 35 26    1    2 0.0132166817 0.6345325786 0.72817011 1.3494860  2.984517   2
#> 36 27    0    1 0.0000000000 0.0008818818 2.16831094 2.1691928 13.508095   2
#> 37 27    1    2 0.0008818818 0.2667189755 2.16919282 2.4350299 13.508095   2
#> 38 28    0    1 0.0000000000 0.4529311732 2.04225701 2.4951882  9.380537   2
#> 39 28    1    2 0.4529311732 0.6866900240 2.49518819 2.7289470  9.380537   2
#> 40 29    0    2 0.0000000000 0.1772835591 1.53481023 1.7120938  9.495196   2
#> 41 30    0    2 0.0000000000 0.1941817385 1.63206007 1.8262418 26.035616   2
#> 42 31    0    1 0.0000000000 0.0308302014 0.35660841 0.3874386 10.551045   2
#> 43 31    1    2 0.0308302014 0.4907436594 0.38743861 0.8473521 10.551045   2
#> 44 32    0    1 0.0000000000 0.0740600793 2.15357481 2.2276349  4.416589   2
#> 45 32    1    2 0.0740600793 0.1101340874 2.22763489 2.2637089  4.416589   2
#> 46 33    0    2 0.0000000000 0.5565792588 1.95045550 2.5070348  7.962846   2
#> 47 34    0    1 0.0000000000 0.0458368413 0.26345926 0.3092961  2.618997   2
#> 48 34    1    2 0.0458368413 0.2517980122 0.30929610 0.5152573  2.618997   2
#> 49 35    0    2 0.0000000000 0.1734901939 0.05170118 0.2251914  5.526203   2
#> 50 36    0    2 0.0000000000 1.2244407660 0.63219546 1.8566362 13.584287   2
#> 51 37    0    2 0.0000000000 0.7366868504 2.82826551 3.5649524 21.853652   2
#> 52 38    0    2 0.0000000000 0.4807117816 1.21027421 1.6909860 15.876691   2
#> 53 39    0    1 0.0000000000 0.4954691576 2.66994815 3.1654173  4.208202   2
#> 54 39    1    2 0.4954691576 0.5979886068 3.16541731 3.2679368  4.208202   2
#> 55 40    0    1 0.0000000000 0.3353758096 2.37495290 2.7103287 25.889694   2
#> 56 40    1    2 0.3353758096 0.5801868097 2.71032871 2.9551397 25.889694   2
#> 
#> [[10]]
#>    id from   to      entry       exit  entryAct   exitAct   censAct trt
#> 1   1    0    2 0.00000000 0.36486251 0.7307464 1.0956089  4.311449   1
#> 2   2    0    2 0.00000000 0.72797665 0.3975393 1.1255159 27.974739   1
#> 3   3    0    2 0.00000000 0.05176356 1.4919859 1.5437495 18.339192   1
#> 4   4    0    1 0.00000000 0.47081912 0.8351729 1.3059921 10.525492   1
#> 5   4    1    2 0.47081912 0.47274349 1.3059921 1.3079164 10.525492   1
#> 6   5    0    1 0.00000000 0.00108168 2.8368515 2.8379332 12.242199   1
#> 7   5    1    2 0.00108168 0.97213778 2.8379332 3.8089893 12.242199   1
#> 8   6    0    2 0.00000000 0.42425732 0.7896894 1.2139467  6.993115   1
#> 9   7    0    2 0.00000000 0.22947801 1.6417992 1.8712772 23.432124   1
#> 10  8    0    1 0.00000000 0.62842411 1.7501951 2.3786192 43.814127   1
#> 11  8    1    2 0.62842411 1.30797638 2.3786192 3.0581715 43.814127   1
#> 12  9    0    1 0.00000000 0.09565990 0.4674800 0.5631399  8.338261   1
#> 13  9    1    2 0.09565990 0.73217961 0.5631399 1.1996596  8.338261   1
#> 14 10    0    1 0.00000000 0.28165594 1.4690742 1.7507301  2.458164   1
#> 15 10    1    2 0.28165594 0.61521628 1.7507301 2.0842904  2.458164   1
#> 16 11    0    2 0.00000000 0.32437014 2.7757701 3.1001402  8.220130   1
#> 17 12    0 cens 0.00000000 0.50272923 1.1150273 1.6177565  1.617757   1
#> 18 13    0    1 0.00000000 0.08175658 0.9442993 1.0260559  2.351395   1
#> 19 13    1    2 0.08175658 0.93597639 1.0260559 1.8802757  2.351395   1
#> 20 14    0    2 0.00000000 0.21832411 2.4607367 2.6790608 18.224221   1
#> 21 15    0    2 0.00000000 0.57085086 0.5809910 1.1518418 24.840938   1
#> 22 16    0    2 0.00000000 0.74476018 2.2087991 2.9535593 14.840138   1
#> 23 17    0    1 0.00000000 0.26544720 0.3033039 0.5687510 19.441766   1
#> 24 17    1    2 0.26544720 0.37534484 0.5687510 0.6786487 19.441766   1
#> 25 18    0    1 0.00000000 0.84428717 1.7122593 2.5565464 31.410514   1
#> 26 18    1    2 0.84428717 0.87849712 2.5565464 2.5907564 31.410514   1
#> 27 19    0    2 0.00000000 0.48067579 0.7149038 1.1955796 20.583463   1
#> 28 20    0    2 0.00000000 0.04306470 1.5771965 1.6202612 36.350843   1
#> 29 21    0    1 0.00000000 0.02016253 1.7171818 1.7373443  9.824775   2
#> 30 21    1    2 0.02016253 0.47052947 1.7373443 2.1877113  9.824775   2
#> 31 22    0    1 0.00000000 0.53639734 1.4663219 2.0027192 10.790572   2
#> 32 22    1    2 0.53639734 0.77034763 2.0027192 2.2366695 10.790572   2
#> 33 23    0    2 0.00000000 0.30599137 1.6790901 1.9850815  4.933431   2
#> 34 24    0    2 0.00000000 0.50274840 0.5745566 1.0773051  4.433750   2
#> 35 25    0    2 0.00000000 0.58849542 2.7953137 3.3838091 26.470819   2
#> 36 26    0    2 0.00000000 0.65169959 1.2948226 1.9465222  8.987117   2
#> 37 27    0    1 0.00000000 0.70027756 2.8019015 3.5021790 13.541888   2
#> 38 27    1    2 0.70027756 1.08742852 3.5021790 3.8893300 13.541888   2
#> 39 28    0    1 0.00000000 0.17277439 0.6548238 0.8275982  3.613415   2
#> 40 28    1    2 0.17277439 0.17844128 0.8275982 0.8332651  3.613415   2
#> 41 29    0    2 0.00000000 0.41708215 1.0315548 1.4486369  7.770879   2
#> 42 30    0    1 0.00000000 0.05264367 0.9757414 1.0283850 12.398919   2
#> 43 30    1    2 0.05264367 0.38904367 1.0283850 1.3647851 12.398919   2
#> 44 31    0    1 0.00000000 0.09280475 1.5675254 1.6603301  1.840354   2
#> 45 31    1 cens 0.09280475 0.27282821 1.6603301 1.8403536  1.840354   2
#> 46 32    0    1 0.00000000 0.41894244 1.0080013 1.4269437 20.215680   2
#> 47 32    1    2 0.41894244 0.68043948 1.4269437 1.6884408 20.215680   2
#> 48 33    0    2 0.00000000 0.39736600 1.6814041 2.0787701 17.231131   2
#> 49 34    0    2 0.00000000 0.02777484 2.5942043 2.6219791 34.867212   2
#> 50 35    0    1 0.00000000 0.52238592 1.2512069 1.7735928 50.060027   2
#> 51 35    1    2 0.52238592 0.59667484 1.7735928 1.8478818 50.060027   2
#> 52 36    0    2 0.00000000 0.58136902 2.1803972 2.7617662  8.645262   2
#> 53 37    0    2 0.00000000 0.01023474 2.4547162 2.4649509  6.553883   2
#> 54 38    0    2 0.00000000 0.61860015 1.1661620 1.7847622 51.228770   2
#> 55 39    0    2 0.00000000 0.19551571 1.6232195 1.8187352 42.039225   2
#> 56 40    0    2 0.00000000 0.01249644 0.3144476 0.3269440 23.696916   2
#>