Helper function for censoringByNumberEvents

Helper function for censoringByNumberEvents

getCensoredData(time, event, data)

Arguments

time

(numeric)
event times.

event

(numeric)
event indicator.

data

(data.frame)
data frame including patient id id, recruiting time recruitTime and individual censoring time censTimeInd.

Value

This function returns a data frame with columns: event time, censoring indicator, event indicator and event time in calendar time.

Examples

transition1 <- weibull_transition(h01 = 1.2, h02 = 1.5, h12 = 1.6, p01 = 0.8, p02 = 0.9, p12 = 1)
transition2 <- weibull_transition(h01 = 1, h02 = 1.3, h12 = 1.7, p01 = 1.1, p02 = 0.9, p12 = 1.1)

simStudy <- getOneClinicalTrial(
  nPat = c(20, 20), transitionByArm = list(transition1, transition2),
  dropout = list(rate = 0.3, time = 10),
  accrual = list(param = "time", value = 7)
)
simStudyWide <- getDatasetWideFormat(simStudy)
simStudyWide$censTimeInd <- 5 - simStudyWide$recruitTime
NotRecruited <- simStudyWide$id[simStudyWide$censTimeInd < 0]
censoredData <- simStudyWide[!(simStudyWide$id %in% NotRecruited), ]
getCensoredData(time = censoredData$OStime, event = censoredData$OSevent, data = censoredData)
#>           time Censored event   timeCal
#> 1  0.095611326        0     1 0.7049806
#> 2  0.128129660        0     1 4.7552251
#> 3  0.456187458        0     1 1.5145144
#> 4  0.617578154        0     1 1.5749634
#> 5  0.941381975        1     0 5.0000000
#> 6  0.464251639        0     1 3.8486554
#> 7  0.201077118        1     0 5.0000000
#> 8  2.514980004        1     0 5.0000000
#> 9  0.355516036        0     1 1.0561903
#> 10 0.438380071        0     1 2.6603980
#> 11 0.044822013        1     0 5.0000000
#> 12 1.292220379        0     1 4.3091421
#> 13 0.006099358        0     1 0.7015168
#> 14 0.318636795        1     0 5.0000000
#> 15 2.472908334        0     1 4.8579634
#> 16 0.857081659        1     0 5.0000000
#> 17 0.642192385        0     1 2.0541822
#> 18 1.230390072        1     0 5.0000000
#> 19 2.791198289        0     1 2.8519452
#> 20 0.001461112        0     1 3.4665408
#> 21 0.182795539        0     1 1.0704293
#> 22 0.594277100        0     1 0.6973992
#> 23 0.471521423        0     1 2.9820266
#> 24 1.814844158        1     0 5.0000000
#> 25 0.749690442        0     1 1.3558039
#> 26 0.852404990        0     1 3.0876434
#> 27 0.035256673        0     1 3.1267463
#> 28 0.621540457        0     1 4.8651038
#> 29 0.275384615        0     1 0.4971745
#> 30 0.397984871        0     1 0.5533860
#> 31 0.451313118        0     1 2.4896527